Store database in ~/.config
The birthdays database is now stored in `~/.config` instead of in the
source tree. This enables the program to point to a single database
regardless of the working directory it is called from, instead of having
it create a new database each time it is called from a working directory
that doesn't include any.